+++ .gitignore
@@ -5,3 +5,4 @@ test.py
 __pycache__
 Researchs
 bbstrader.egg-info
+WS.code-workspace
